Coordinates still searching out of bounds.

+++ Template/textures.hpp
@@ -49,8 +49,8 @@ namespace graphics {
 			return m_out_texture[(y % m_height) * m_width + (x % m_width)]; // Finds correct coord, loops if necessary
 		}
 
-		int64 width() { return m_width; }
-		int64 height() { return m_height; }
+		float width()      { return float(m_width); }
+		float height()     { return float(m_height); }
 		ivec2 dimensions() { return m_dimensions; }
 	};
